+++ substrate/frame/support/test/tests/construct_runtime_ui/pallet_error_too_large.rs
@@ -54,7 +54,7 @@ pub enum Nested4 {
 }
 
 pub type Signature = sr25519::Signature;
-pub type BlockNumber = u32;
+pub type BlockNumber = u64;
 pub type Header = generic::Header<BlockNumber, BlakeTwo256>;
 pub type Block = generic::Block<Header, UncheckedExtrinsic>;
 pub type UncheckedExtrinsic = generic::UncheckedExtrinsic<u32, RuntimeCall, Signature, ()>;
