+++ substrate/frame/support/test/tests/storage_transaction.rs
@@ -72,7 +72,7 @@ pub mod pallet {
 	pub type Map<T: Config> = StorageMap<_, Twox64Concat, String, u32, ValueQuery>;
 }
 
-pub type BlockNumber = u32;
+pub type BlockNumber = u64;
 pub type Signature = sr25519::Signature;
 pub type AccountId = <Signature as Verify>::Signer;
 pub type Header = generic::Header<BlockNumber, BlakeTwo256>;
